Moved IActivationFunction and related types to from the SharpNeat.NeuralNets namespace to SharpNeat.NeuralNets.ActivationFunctions
