+++ src/SharpNeat/Graphs/IActivationFunctionLibrary.cs
@@ -1,7 +1,7 @@
 // This file is part of SharpNEAT; Copyright Colin D. Green.
 // See LICENSE.txt for details.
 using System.Numerics;
-using SharpNeat.NeuralNets;
+using SharpNeat.NeuralNets.ActivationFunctions;
 
 namespace SharpNeat.Graphs;
