+++ src/SharpNeat/NeuralNets/Double/Vectorized/NeuralNetAcyclic.cs
@@ -6,6 +6,7 @@ using System.Numerics;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 using SharpNeat.Graphs.Acyclic;
+using SharpNeat.NeuralNets.ActivationFunctions;
 
 namespace SharpNeat.NeuralNets.Double.Vectorized;
