+++ src/SharpNeat/NeuralNets/Double/Vectorized/NeuralNetCyclic.cs
@@ -5,6 +5,7 @@ using System.Diagnostics;
 using System.Numerics;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
+using SharpNeat.NeuralNets.ActivationFunctions;
 
 namespace SharpNeat.NeuralNets.Double.Vectorized;
