+++ src/Tests/SharpNeat.Tests/NeuralNets/Double/ActivationFunctionFactoryTests.cs
@@ -1,4 +1,5 @@
-﻿using Xunit;
+﻿using SharpNeat.NeuralNets.ActivationFunctions;
+using Xunit;
 
 namespace SharpNeat.NeuralNets.Double;
